Moved Steam manifest to launch dependency
